requirements.txt
# Visualization packages
bokeh
seaborn
matplotlib

#pygal
#Meta learning
auto-sklearn
optunity
# Vision
scikit-image

#Satellite imagery
descartes

# Stats Packages
statsmodels

# Feature selector for timeseries
tsfresh

# Genetic algo based feature selection
# Tools for tpot
deap
update_checker
tqdm
tpot

# notebook tools
jupyter
psycopg2

# matrix tool + data structure/frame
numpy
pandas
sqlalchemy
scikit-learn
scipy
scikit-plot

#Non-linear correlation measures
ace

#Xtreme Gradient Boosting
xgboost

#Rasterio reads and writes geospatial raster datasets
rasterio

#Light GBM
-e git+https://github.com/Microsoft/LightGBM.git#egg=lightgbm
-e git+https://github.com/ArdalanM/pyLightGBM.git#egg=pylightgbm


# Package for map projections
geos

# Neural network packages
gym
keras
mxnet
-e git+https://github.com:greytip/SOMPY.git#egg=sompy

# Vision packages
#object detections
gpod

#Workflow management

#dagpype
#luigi
